// Epitech/JobBoard/backoffice/src/enum/logType.enum.ts
export enum LogType {
  ADMIN_UPDATE_USER = 'admin.update.user',

  USER_REGISTER = 'user.register',
  USER_FORGOT_PASSWORD = 'user.forgot_password',
  USER_UPDATE_INFORMATION = 'user.update.information',
  USER_UPDATE_PASSWORD = 'user.update.password',
  USER_UPDATE_AVATAR = 'user.update.avatar',
  USER_UPDATE_CURRICULUM = 'user.update.curriculumVitae',
  USER_DELETE = 'user.delete',

  COMPANY_CREATE = 'company.create',
  COMPANY_ADD_USER = 'company.add.user',
  COMPANY_DELETE = 'company.delete',

  ADVERTISMENT_APPLY = 'advertisment.apply',
  ADVERTISMENT_DELETE = 'advertisment.delete',
}